rename metaMarket.ts to hit.ts
